test parallels with all channel types also instead of only the hardcoded default
